Pycalver2

Closes #8, #12, #2, #11, #10, #9, and #7

See merge request mbarkhau/pycalver!4

+++ src/bumpver/vcs.py
@@ -1,12 +1,13 @@
 # This file is part of the pycalver project
-# https://gitlab.com/mbarkhau/pycalver
+# https://github.com/mbarkhau/pycalver
 #
-# Copyright (c) 2019 Manuel Barkhau (mbarkhau@gmail.com) - MIT License
+# Copyright (c) 2018-2020 Manuel Barkhau (mbarkhau@gmail.com) - MIT License
 # SPDX-License-Identifier: MIT
 #
-# pycalver/vcs.py (this file) is based on code from the
+# bumpver/vcs.py (this file) is based on code from the
 # bumpversion project: https://github.com/peritus/bumpversion
 # Copyright (c) 2013-2014 Filip Noetzel - MIT License
+
 """Minimal Git and Mercirial API.
 
 If terminology for similar concepts differs between git and
@@ -15,12 +16,16 @@ mercurial, then the git terms are used. For example "fetch"
 """
 
 import os
+import sys
+import shlex
 import typing as typ
 import logging
 import tempfile
 import subprocess as sp
 
-logger = logging.getLogger("pycalver.vcs")
+from . import config
+
+logger = logging.getLogger("bumpver.vcs")
 
 
 VCS_SUBCOMMANDS_BY_NAME = {
@@ -30,9 +35,9 @@ VCS_SUBCOMMANDS_BY_NAME = {
         'ls_tags'     : "git tag --list",
         'status'      : "git status --porcelain",
         'add_path'    : "git add --update {path}",
-        'commit'      : "git commit --file {path}",
+        'commit'      : "git commit --message '{message}'",
         'tag'         : "git tag --annotate {tag} --message {tag}",
-        'push_tag'    : "git push origin --follow-tags {tag}",
+        'push_tag'    : "git push origin --follow-tags {tag} HEAD",
         'show_remotes': "git config --get remote.origin.url",
     },
     'hg': {
@@ -70,11 +75,10 @@ class VCSAPI:
             logger.info(cmd_str)
         else:
             logger.debug(cmd_str)
-        output_data: bytes = sp.check_output(cmd_str.split(), env=env, stderr=sp.STDOUT)
+        cmd_parts = shlex.split(cmd_str)
+        output_data: bytes = sp.check_output(cmd_parts, env=env, stderr=sp.STDOUT)
 
-        # TODO (mb 2018-11-15): Detect encoding of output?
-        _encoding = "utf-8"
-        return output_data.decode(_encoding)
+        return output_data.decode("utf-8")
 
     @property
     def is_usable(self) -> bool:
@@ -96,11 +100,13 @@ class VCSAPI:
 
     @property
     def has_remote(self) -> bool:
+        # pylint:disable=broad-except;  Not sure how to anticipate all cases.
         try:
             output = self('show_remotes')
             if output.strip() == "":
                 return False
-            return True
+            else:
+                return True
         except Exception:
             return False
 
@@ -139,20 +145,25 @@ class VCSAPI:
 
     def commit(self, message: str) -> None:
         """Commit added files."""
-        message_data = message.encode("utf-8")
-
-        tmp_file = tempfile.NamedTemporaryFile("wb", delete=False)
-        assert " " not in tmp_file.name
-
-        fobj: typ.IO[bytes]
-
-        with tmp_file as fobj:
-            fobj.write(message_data)
-
         env: Env = os.environ.copy()
-        env['HGENCODING'] = "utf-8"
-        self('commit', env=env, path=tmp_file.name)
-        os.unlink(tmp_file.name)
+
+        if self.name == 'git':
+            self('commit', env=env, message=message)
+        else:
+            message_data = message.encode("utf-8")
+            tmp_file     = tempfile.NamedTemporaryFile("wb", delete=False)
+            try:
+                assert " " not in tmp_file.name
+
+                fobj: typ.IO[bytes]
+
+                with tmp_file as fobj:
+                    fobj.write(message_data)
+
+                env['HGENCODING'] = "utf-8"
+                self('commit', env=env, path=tmp_file.name)
+            finally:
+                os.unlink(tmp_file.name)
 
     def tag(self, tag_name: str) -> None:
         """Create an annotated tag."""
@@ -179,3 +190,57 @@ def get_vcs_api() -> VCSAPI:
             return vcs_api
 
     raise OSError("No such directory .git/ or .hg/ ")
+
+
+# cli helper methods
+
+
+def assert_not_dirty(vcs_api: VCSAPI, filepaths: typ.Set[str], allow_dirty: bool) -> None:
+    dirty_files = vcs_api.status(required_files=filepaths)
+
+    if dirty_files:
+        logger.warning(f"{vcs_api.name} working directory is not clean. Uncomitted file(s):")
+        for dirty_file in dirty_files:
+            logger.warning("    " + dirty_file)
+
+    if not allow_dirty and dirty_files:
+        sys.exit(1)
+
+    dirty_pattern_files = set(dirty_files) & filepaths
+    if dirty_pattern_files:
+        logger.error("Not commiting when pattern files are dirty:")
+        for dirty_file in dirty_pattern_files:
+            logger.warning("    " + dirty_file)
+        sys.exit(1)
+
+
+def commit(
+    cfg           : config.Config,
+    vcs_api       : VCSAPI,
+    filepaths     : typ.Set[str],
+    new_version   : str,
+    commit_message: str,
+) -> None:
+    for filepath in filepaths:
+        vcs_api.add(filepath)
+
+    vcs_api.commit(commit_message)
+
+    if cfg.commit and cfg.tag:
+        vcs_api.tag(new_version)
+
+    if cfg.commit and cfg.tag and cfg.push:
+        vcs_api.push(new_version)
+
+
+def get_tags(fetch: bool) -> typ.List[str]:
+    try:
+        vcs_api = get_vcs_api()
+        logger.debug(f"vcs found: {vcs_api.name}")
+        if fetch:
+            logger.info("fetching tags from remote (to turn off use: -n / --no-fetch)")
+            vcs_api.fetch()
+        return vcs_api.ls_tags()
+    except OSError:
+        logger.debug("No vcs found")
+        return []
